added client part

+++ app/src/main/java/com/kotleters/mobile/common/domain/Company.kt
@@ -10,7 +10,7 @@ data class Company(
     data class Offer(
         val title: String,
         val description: String,
-        val discount: Long,
+        val discount: Double,
         val startDate: LocalDateTime,
         val endDate: LocalDateTime
     )
